refacto(validator): move checkControlFilePerm to utils(_test)
